+++ src/apps/obsidian-plugin/views/RightSidebarReactView/ScheduledItemsSection/Tabs/CalendarItemsList.tsx
@@ -17,7 +17,7 @@ import {
 } from "@mui/material";
 import { ResponsiveScheduledItem } from "apps/obsidian-plugin/components/ResponsiveScheduledItem";
 import { useLogger } from "apps/obsidian-plugin/hooks";
-import { EditItemRecurrencePanel } from "apps/obsidian-plugin/panels/CreateBudgetItemPanel/EditItemRecurrencePanel";
+import { EditScheduledTransactionPanel } from "apps/obsidian-plugin/panels/CreateBudgetItemPanel/EditScheduledTransactionPanel";
 import { RecordItemPanel } from "apps/obsidian-plugin/panels/RecordItemPanel";
 import { AccountBalance, AccountName } from "contexts/Accounts/domain";
 import { CategoryID } from "contexts/Categories/domain";
@@ -958,17 +958,16 @@ const CalendarItemsListItem = ({
 					recurrence.date.value.getTime() && (
 					<>
 						{showPanel.action === "edit" && (
-							<EditItemRecurrencePanel
+							<EditScheduledTransactionPanel
 								scheduledTransaction={scheduledTransaction}
-								recurrence={{
-									recurrence: showPanel.item.recurrence,
-								}}
+								recurrence={showPanel.item.recurrence}
 								onClose={() => {
 									setShowPanel(undefined);
 									setSelectedItem(undefined);
 									setAction(undefined);
 								}}
 								updateItems={updateItems}
+								initialScope="single"
 							/>
 						)}
 						{showPanel.action === "record" && (